1. 調整 Meats 命名空間下的類別移動至 Models.Meats 命名空間下，並調整相關參考的程式碼。 2. 格式化與清理 Meats.cs 文件並增加 XML 註解。 3. 調整 Meats 為抽象類別及 Meats.Name 屬性為 protected set。 4. 格式化與清理 Bacon.cs 文件並增加 XML 註解。 5. 格式化與清理 TaiwanChicken.cs 文件並增加 XML 註解。 6. 格式化與清理 Tenderloin.cs 文件並增加 XML 註解。 7. 格式化與清理 Turkey.cs 文件並增加 XML 註解。
