--- NineYi.BurgerShop/Meats/Bacon.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Meats
-{
-    /// <summary>
-    /// 培根
-    /// </summary>
-    public class Bacon : Meat
-    {
-        public Bacon()
-        {
-            this.Name = "Bacon";
-        }
-    }
-}
