--- NineYi.BurgerShop/Meats/Meat.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Meats
-{
-    /// <summary>
-    /// 肉類
-    /// </summary>
-    public class Meat
-    {
-        public string Name { get; set; }
-
-        public override string ToString()
-        {
-            return this.Name;
-        }
-    }
-}
