--- NineYi.BurgerShop/Meats/TaiwanChicken.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Meats
-{
-    /// <summary>
-    /// 台灣土雞
-    /// </summary>
-    public class TaiwanChicken : Meat
-    {
-        public TaiwanChicken()
-        {
-            this.Name = "Taiwan Chicken";
-        }
-    }
-}
