--- NineYi.BurgerShop/Meats/Tenderloin.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Meats
-{
-    /// <summary>
-    /// 豬里肌肉
-    /// </summary>
-    public class Tenderloin : Meat
-    {
-        public Tenderloin()
-        {
-            this.Name = "Tenderloin";
-        }
-    }
-}
