--- NineYi.BurgerShop/Meats/Turkey.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Meats
-{
-    /// <summary>
-    /// 火雞
-    /// </summary>
-    public class Turkey : Meat
-    {
-        public Turkey()
-        {
-            this.Name = "Turkey";
-        }
-    }
-}
