+++ NineYi.BurgerShop/Models/Meats/Bacon.cs
@@ -1,0 +1,17 @@
+﻿namespace NineYi.BurgerShop.Models.Meats
+{
+    /// <summary>
+    /// 培根。
+    /// </summary>
+    /// <seealso cref="Meat" />
+    public class Bacon : Meat
+    {
+        /// <summary>
+        /// 初始化 <see cref="Bacon"/> 類別新的執行個體。
+        /// </summary>
+        public Bacon()
+        {
+            this.Name = "Bacon";
+        }
+    }
+}
