+++ NineYi.BurgerShop/Models/Meats/Tenderloin.cs
@@ -1,0 +1,17 @@
+﻿namespace NineYi.BurgerShop.Models.Meats
+{
+    /// <summary>
+    /// 豬里肌肉。
+    /// </summary>
+    /// <seealso cref="NineYi.BurgerShop.Models.Meats.Meat" />
+    public class Tenderloin : Meat
+    {
+        /// <summary>
+        /// 初始化 <see cref="Tenderloin"/> 類別新的執行個體。
+        /// </summary>
+        public Tenderloin()
+        {
+            this.Name = "Tenderloin";
+        }
+    }
+}
