+++ NineYi.BurgerShop/Models/Meats/Turkey.cs
@@ -1,0 +1,17 @@
+﻿namespace NineYi.BurgerShop.Models.Meats
+{
+    /// <summary>
+    /// 火雞。
+    /// </summary>
+    /// <seealso cref="NineYi.BurgerShop.Models.Meats.Meat" />
+    public class Turkey : Meat
+    {
+        /// <summary>
+        /// 初始化 <see cref="Turkey"/> 類別新的執行個體。
+        /// </summary>
+        public Turkey()
+        {
+            this.Name = "Turkey";
+        }
+    }
+}
